telegramBot/requirements.txt
aiogram~=3.17.0
aiofiles~=24.1.0
python-dotenv~=1.0.1
pydantic~=2.10.6